Internal: Move ExerciseAccessHelperTrait to the shared Traits namespace
